// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "posts".
 */
export interface Post {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  coverImage?: (number | null) | Media;
  content?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  status?: ('draft' | 'published') | null;
  publishedAt?: string | null;
  tags?:
    | {
        tag?: string | null;
        id?: string | null;
      }[]
    | null;
  /**
   * Estimated read time in minutes
   */
  readTime?: number | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "projects".
 */
export interface Project {
  id: number;
  title: string;
  slug: string;
  description: string;
  longDescription?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  coverImage?: (number | null) | Media;
  screenshots?:
    | {
        image?: (number | null) | Media;
        caption?: string | null;
        id?: string | null;
      }[]
    | null;
  techStack?:
    | {
        tech?: string | null;
        id?: string | null;
      }[]
    | null;
  githubUrl?: string | null;
  liveUrl?: string | null;
  featured?: boolean | null;
  status?: ('active' | 'archived' | 'in-progress') | null;
  category?: ('web' | 'mobile' | 'cli' | 'oss' | 'other') | null;
  order?: number | null;
  updatedAt: string;
  createdAt: string;
}